TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/pop3/command/PASSCommandTest.java
package de.rub.nds.tlsattacker.core.pop3.command;

import de.rub.nds.tlsattacker.core.connection.OutboundConnection;
import de.rub.nds.tlsattacker.core.layer.context.Pop3Context;
import de.rub.nds.tlsattacker.core.pop3.parser.command.Pop3CommandParser;
import de.rub.nds.tlsattacker.core.pop3.parser.command.Pop3PASSCommandParser;
import de.rub.nds.tlsattacker.core.state.Context;
import de.rub.nds.tlsattacker.core.state.State;
import org.junit.jupiter.api.Test;

import java.io.ByteArrayInputStream;
import java.nio.charset.StandardCharsets;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class PASSCommandTest {

    @Test
    void testParse() {
        Pop3Context context = new Pop3Context(new Context(new State(), new OutboundConnection()));
        PASSCommand passCommand = new PASSCommand();
        String message = "PASS p4ssw0rd\r\n";

        Pop3PASSCommandParser parser = passCommand.getParser(context, new ByteArrayInputStream(message.getBytes(StandardCharsets.UTF_8)));
        parser.parse(passCommand);

        assertEquals(passCommand.getPassword(), "p4ssw0rd");
        assertEquals(passCommand.getCommandName(), "PASS");
    }
}
